core: Extract interactions communication
