+++ src/core/bonded_interactions/angle_cossquare.cpp
@@ -24,7 +24,7 @@
  */
 #include "angle_cossquare.hpp"
 
-#include "communication.hpp"
+#include "interactions.hpp"
 
 int angle_cossquare_set_params(int bond_type, double bend, double phi0) {
   if (bond_type < 0)
