+++ src/core/bonded_interactions/angle_harmonic.cpp
@@ -24,7 +24,7 @@
  */
 #include "angle_harmonic.hpp"
 
-#include "communication.hpp"
+#include "interactions.hpp"
 
 #include <utils/constants.hpp>
