+++ src/core/bonded_interactions/bonded_coulomb.cpp
@@ -23,7 +23,7 @@
  *  Implementation of \ref bonded_coulomb.hpp
  */
 #include "bonded_coulomb.hpp"
-#include "communication.hpp"
+#include "interactions.hpp"
 
 #include <utils/constants.hpp>
