+++ src/core/bonded_interactions/fene.cpp
@@ -24,7 +24,7 @@
  */
 
 #include "fene.hpp"
-#include "communication.hpp"
+#include "interactions.hpp"
 
 #include <utils/constants.hpp>
 #include <utils/math/sqr.hpp>
