+++ src/core/bonded_interactions/harmonic.cpp
@@ -23,7 +23,7 @@
  *  Implementation of \ref harmonic.hpp
  */
 #include "harmonic.hpp"
-#include "communication.hpp"
+#include "interactions.hpp"
 
 #include <utils/constants.hpp>
