+++ src/core/bonded_interactions/harmonic_dumbbell.cpp
@@ -25,7 +25,7 @@
 #include "harmonic_dumbbell.hpp"
 
 #ifdef ROTATION
-#include "communication.hpp"
+#include "interactions.hpp"
 
 #include <utils/constants.hpp>
