+++ src/core/immersed_boundary/ImmersedBoundaries.cpp
@@ -23,6 +23,7 @@
 #include "bonded_interactions/bonded_interaction_data.hpp"
 #include "communication.hpp"
 #include "grid.hpp"
+#include "interactions.hpp"
 
 #include <utils/constants.hpp>
