+++ src/core/nonbonded_interactions/buckingham.cpp
@@ -25,7 +25,7 @@
 #include "buckingham.hpp"
 
 #ifdef BUCKINGHAM
-#include "communication.hpp"
+#include "interactions.hpp"
 
 #include <utils/constants.hpp>
