+++ src/core/nonbonded_interactions/thole.cpp
@@ -25,7 +25,7 @@
 #include "thole.hpp"
 
 #ifdef THOLE
-#include "communication.hpp"
+#include "interactions.hpp"
 
 #include <utils/constants.hpp>
