+++ src/core/nonbonded_interactions/wca.cpp
@@ -23,7 +23,7 @@
 #include "wca.hpp"
 
 #ifdef WCA
-#include "communication.hpp"
+#include "interactions.hpp"
 
 #include <utils/constants.hpp>
